Refactoring of some tests to separate robolectric dependent tests from tests not dependant on robolectric.
