+++ libnavui-maps/src/test/java/com/mapbox/navigation/ui/maps/route/line/api/VanishingRouteLineTest.kt
@@ -1,18 +1,9 @@
 package com.mapbox.navigation.ui.maps.route.line.api
 
-import android.graphics.Color
-import com.mapbox.api.directions.v5.models.DirectionsRoute
-import com.mapbox.core.constants.Constants
-import com.mapbox.geojson.LineString
-import com.mapbox.geojson.Point
 import com.mapbox.navigation.base.trip.model.RouteProgressState
-import com.mapbox.navigation.testing.FileUtils.loadJsonFixture
 import com.mapbox.navigation.testing.MainCoroutineRule
-import com.mapbox.navigation.ui.maps.internal.route.line.MapboxRouteLineUtils
-import com.mapbox.navigation.ui.maps.route.line.model.RouteLineColorResources
-import com.mapbox.navigation.ui.maps.route.line.model.RouteLineExpressionData
-import com.mapbox.navigation.ui.maps.route.line.model.RouteLineResources
 import com.mapbox.navigation.ui.maps.route.line.model.VanishingPointState
+import com.mapbox.navigation.ui.maps.testing.TestingUtil.loadRoute
 import com.mapbox.navigation.utils.internal.InternalJobControlFactory
 import com.mapbox.navigation.utils.internal.JobControl
 import io.mockk.every
@@ -22,7 +13,6 @@ import io.mockk.unmockkObject
 import io.mockk.verify
 import kotlinx.coroutines.CompletableJob
 import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.SupervisorJob
 import kotlinx.coroutines.cancelChildren
 import org.junit.After
@@ -32,11 +22,7 @@ import org.junit.Assert.assertNull
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
-import org.junit.runner.RunWith
-import org.robolectric.RobolectricTestRunner
 
-@ExperimentalCoroutinesApi
-@RunWith(RobolectricTestRunner::class)
 class VanishingRouteLineTest {
 
     @get:Rule
@@ -57,47 +43,6 @@ class VanishingRouteLineTest {
     @After
     fun cleanUp() {
         unmockkObject(InternalJobControlFactory)
-    }
-
-    @Test
-    fun initWithRoute() = coroutineRule.runBlockingTest {
-        val route = getRoute()
-
-        val vanishingRouteLine = VanishingRouteLine().also {
-            it.initWithRoute(route)
-        }
-
-        assertEquals(
-            0.0000025451727518618744,
-            vanishingRouteLine.primaryRouteLineGranularDistances!!.distance,
-            0.0
-        )
-        assertEquals(
-            8,
-            vanishingRouteLine.primaryRouteLineGranularDistances!!.distancesArray.size()
-        )
-    }
-
-    @Test
-    fun initWithRoute_clearState() = coroutineRule.runBlockingTest {
-        val route = getRoute()
-        val vanishingRouteLine = VanishingRouteLine()
-        vanishingRouteLine.setJobControl(testJobControl)
-        vanishingRouteLine.initWithRoute(route)
-        assertEquals(
-            8,
-            vanishingRouteLine.primaryRouteLineGranularDistances!!.distancesArray.size()
-        )
-        unmockkObject(InternalJobControlFactory)
-        val jobControl = InternalJobControlFactory.createDefaultScopeJobControl()
-        vanishingRouteLine.setJobControl(jobControl)
-
-        Runnable {
-            vanishingRouteLine.initWithRoute(getVeryLongRoute())
-        }.run()
-        jobControl.job.cancel()
-
-        assertNull(vanishingRouteLine.primaryRouteLineGranularDistances)
     }
 
     @Test
@@ -133,7 +78,7 @@ class VanishingRouteLineTest {
     @Test
     fun clear() = coroutineRule.runBlockingTest {
         val vanishingRouteLine = VanishingRouteLine().also {
-            it.initWithRoute(getRoute())
+            it.initWithRoute(loadRoute("short_route.json"))
         }
         assertNotNull(vanishingRouteLine.primaryRoutePoints)
         assertNotNull(vanishingRouteLine.primaryRouteLineGranularDistances)
@@ -142,207 +87,6 @@ class VanishingRouteLineTest {
 
         assertNull(vanishingRouteLine.primaryRoutePoints)
         assertNull(vanishingRouteLine.primaryRouteLineGranularDistances)
-    }
-
-    @Test
-    fun getTraveledRouteLineExpressions() {
-        val expectedTrafficExpression = "[step, [line-progress], [rgba, 0.0, 0.0, 0.0, 0.0], 0.0," +
-            " [rgba, 255.0, 255.0, 255.0, 1.0]]"
-        val expectedRouteLineExpression = "[step, [line-progress], [rgba, 0.0, 0.0, 0.0, 0.0]," +
-            " 0.0, [rgba, 0.0, 0.0, 3.0, 0.0]]"
-        val expectedCasingExpression = "[step, [line-progress], [rgba, 0.0, 0.0, 10.0, 0.0]," +
-            " 0.0, [rgba, 0.0, 0.0, 4.0, 0.0]]"
-
-        val colorResources = RouteLineColorResources.Builder()
-            .routeModerateCongestionColor(-1)
-            .routeUnknownCongestionColor(-1)
-            .build()
-
-        val route = getRoute()
-        val lineString = LineString.fromPolyline(route.geometry() ?: "", Constants.PRECISION_6)
-        val vanishingRouteLine = VanishingRouteLine()
-        vanishingRouteLine.initWithRoute(route)
-        vanishingRouteLine.primaryRouteRemainingDistancesIndex = 1
-        val segments: List<RouteLineExpressionData> =
-            MapboxRouteLineUtils.calculateRouteLineSegments(
-                getRoute(),
-                listOf(),
-                true,
-                colorResources
-            )
-
-        val result = vanishingRouteLine.getTraveledRouteLineExpressions(
-            lineString.coordinates()[0],
-            segments,
-            null,
-            genericMockResourceProvider,
-            -1,
-            0.0,
-            false
-        )
-
-        assertEquals(
-            expectedTrafficExpression,
-            result!!.trafficLineExpression.generateExpression().toString()
-        )
-        assertEquals(
-            expectedRouteLineExpression,
-            result.routeLineExpression.generateExpression().toString()
-        )
-        assertEquals(
-            expectedCasingExpression,
-            result.routeLineCasingExpression.generateExpression().toString()
-        )
-    }
-
-    @Test
-    fun getTraveledRouteLineExpressions_withRestrictedLineExpressionData() {
-        val expectedRestrictedExpression = "[step, [line-progress], [rgba, 0.0, 0.0, 0.0, 0.0]," +
-            " 0.0, [rgba, 0.0, 0.0, 0.0, 0.0], 0.44865144220494346, [rgba, 0.0, 0.0, 12.0, 0.0]," +
-            " 0.468779750455607, [rgba, 0.0, 0.0, 0.0, 0.0], 0.5032854217424586, " +
-            "[rgba, 0.0, 0.0, 12.0, 0.0], 0.5207714038134984, [rgba, 0.0, 0.0, 0.0, 0.0]]"
-
-        val colorResources = RouteLineColorResources.Builder()
-            .restrictedRoadColor(Color.CYAN)
-            .build()
-
-        val route = loadRoute("route-with-restrictions.json")
-        val lineString = LineString.fromPolyline(route.geometry() ?: "", Constants.PRECISION_6)
-        val vanishingRouteLine = VanishingRouteLine()
-        vanishingRouteLine.initWithRoute(route)
-        vanishingRouteLine.primaryRouteRemainingDistancesIndex = 1
-        val segments: List<RouteLineExpressionData> =
-            MapboxRouteLineUtils.calculateRouteLineSegments(
-                getRoute(),
-                listOf(),
-                true,
-                colorResources
-            )
-        val restrictedSegments = MapboxRouteLineUtils.extractRouteData(
-            route,
-            MapboxRouteLineUtils.getRouteLegTrafficCongestionProvider
-        )
-
-        val result = vanishingRouteLine.getTraveledRouteLineExpressions(
-            lineString.coordinates()[0],
-            segments,
-            restrictedSegments,
-            genericMockResourceProvider,
-            0,
-            0.0,
-            false
-        )
-
-        assertEquals(
-            expectedRestrictedExpression,
-            result!!.restrictedRoadExpression!!.generateExpression().toString()
-        )
-    }
-
-    @Test
-    fun getTraveledRouteLineExpressions_whenSoftGradientTrue() {
-        val expectedTrafficExpression = "[interpolate, [linear], [line-progress], 0.0, " +
-            "[rgba, 255.0, 255.0, 255.0, 1.0]]"
-        val expectedRouteLineExpression = "[step, [line-progress], [rgba, 0.0, 0.0, 0.0, 0.0]," +
-            " 0.0, [rgba, 0.0, 0.0, 3.0, 0.0]]"
-        val expectedCasingExpression = "[step, [line-progress], [rgba, 0.0, 0.0, 10.0, 0.0]," +
-            " 0.0, [rgba, 0.0, 0.0, 4.0, 0.0]]"
-
-        val colorResources = RouteLineColorResources.Builder()
-            .routeModerateCongestionColor(-1)
-            .routeUnknownCongestionColor(-1)
-            .build()
-
-        val route = getRoute()
-        val lineString = LineString.fromPolyline(route.geometry() ?: "", Constants.PRECISION_6)
-        val vanishingRouteLine = VanishingRouteLine()
-        vanishingRouteLine.initWithRoute(route)
-        vanishingRouteLine.primaryRouteRemainingDistancesIndex = 1
-        val segments: List<RouteLineExpressionData> =
-            MapboxRouteLineUtils.calculateRouteLineSegments(
-                getRoute(),
-                listOf(),
-                true,
-                colorResources
-            )
-
-        val result = vanishingRouteLine.getTraveledRouteLineExpressions(
-            lineString.coordinates()[0],
-            segments,
-            null,
-            genericMockResourceProvider,
-            -1,
-            20.0,
-            true
-        )
-
-        assertEquals(
-            expectedTrafficExpression,
-            result!!.trafficLineExpression.generateExpression().toString()
-        )
-        assertEquals(
-            expectedRouteLineExpression,
-            result.routeLineExpression.generateExpression().toString()
-        )
-        assertEquals(
-            expectedCasingExpression,
-            result.routeLineCasingExpression.generateExpression().toString()
-        )
-    }
-
-    @Test
-    fun getTraveledRouteLineExpressions_multilegRoute_deEmphasizeNonActiveLegs() {
-        val expectedTrafficExp = "[step, [line-progress], [rgba, 0.0, 0.0, 0.0, 0.0], " +
-            "0.13240839439705454, [rgba, 86.0, 168.0, 251.0, 1.0], 0.18199725933538882, " +
-            "[rgba, 255.0, 149.0, 0.0, 1.0], 0.2256358178763112, [rgba, 86.0, 168.0, 251.0, 1.0]," +
-            " 0.32147751186805656, [rgba, 255.0, 149.0, 0.0, 1.0], 0.3838765722116185, " +
-            "[rgba, 86.0, 168.0, 251.0, 1.0], 0.4891841628737826, [rgba, 0.0, 0.0, 0.0, 0.0]]"
-        val expectedRouteLineExp = "[step, [line-progress], [rgba, 0.0, 0.0, 0.0, 0.0], " +
-            "0.13240839439705454, [rgba, 86.0, 168.0, 251.0, 1.0], 0.4891841628737826, " +
-            "[rgba, 0.0, 0.0, 0.0, 0.0]]"
-        val expectedCasingExp = "[step, [line-progress], [rgba, 0.0, 0.0, 0.0, 0.0]," +
-            " 0.13240839439705454, [rgba, 47.0, 122.0, 198.0, 1.0], 0.4891841628737826," +
-            " [rgba, 0.0, 0.0, 0.0, 0.0]]"
-        val route = getMultilegWithTwoLegs()
-        val vanishingRouteLine = VanishingRouteLine()
-        vanishingRouteLine.initWithRoute(route)
-        vanishingRouteLine.primaryRouteRemainingDistancesIndex = 7
-        vanishingRouteLine.vanishPointOffset = 0.1322571610688955
-        val segments = listOf(
-            RouteLineExpressionData(0.0, -27392, 0),
-            RouteLineExpressionData(0.07932429566550842, -45747, 0),
-            RouteLineExpressionData(0.10338667841237215, -11097861, 0),
-            RouteLineExpressionData(0.18199725933538882, -27392, 0),
-            RouteLineExpressionData(0.2256358178763112, -11097861, 0),
-            RouteLineExpressionData(0.32147751186805656, -27392, 0),
-            RouteLineExpressionData(0.3838765722116185, -11097861, 0),
-            RouteLineExpressionData(0.4891841628737826, 0, 1),
-            RouteLineExpressionData(0.5402820600662328, 0, 1),
-            RouteLineExpressionData(0.9738127865054893, 0, 1)
-        )
-
-        val result = vanishingRouteLine.getTraveledRouteLineExpressions(
-            Point.fromLngLat(-122.52351984901476, 37.97384101461195),
-            segments,
-            null,
-            RouteLineResources.Builder().build(),
-            0,
-            0.0,
-            false
-        )
-
-        assertEquals(
-            expectedTrafficExp,
-            result!!.trafficLineExpression.generateExpression().toString()
-        )
-        assertEquals(
-            expectedRouteLineExp,
-            result.routeLineExpression.generateExpression().toString()
-        )
-        assertEquals(
-            expectedCasingExp,
-            result.routeLineCasingExpression.generateExpression().toString()
-        )
     }
 
     @Test
@@ -356,41 +100,5 @@ class VanishingRouteLineTest {
         VanishingRouteLine().cancel()
 
         verify { mockParentJob.cancelChildren() }
-    }
-
-    private fun getRoute(): DirectionsRoute {
-        return loadRoute("short_route.json")
-    }
-
-    private fun getVeryLongRoute(): DirectionsRoute {
-        return loadRoute("cross-country-route.json")
-    }
-
-    private fun loadRoute(routeFileName: String): DirectionsRoute {
-        val routeAsJson = loadJsonFixture(routeFileName)
-        return DirectionsRoute.fromJson(routeAsJson)
-    }
-
-    private val genericMockResourceProvider = mockk<RouteLineResources> {
-        every { routeLineColorResources } returns mockk<RouteLineColorResources> {
-            every { routeUnknownCongestionColor } returns 1
-            every { routeLineTraveledColor } returns 0
-            every { routeDefaultColor } returns 3
-            every { routeCasingColor } returns 4
-            every { routeLowCongestionColor } returns 5
-            every { routeModerateCongestionColor } returns 6
-            every { routeSevereCongestionColor } returns 7
-            every { routeHeavyCongestionColor } returns 8
-            every { alternativeRouteUnknownCongestionColor } returns 9
-            every { routeLineTraveledCasingColor } returns 10
-            every { inActiveRouteLegsColor } returns 11
-            every { restrictedRoadColor } returns 12
-        }
-        every { trafficBackfillRoadClasses } returns listOf()
-    }
-
-    private fun getMultilegWithTwoLegs(): DirectionsRoute {
-        val routeAsJson = loadJsonFixture("multileg-route-two-legs.json")
-        return DirectionsRoute.fromJson(routeAsJson)
     }
 }
