+++ libnavui-maps/src/test/java/com/mapbox/navigation/ui/maps/testing/TestingUtil.kt
@@ -1,0 +1,11 @@
+package com.mapbox.navigation.ui.maps.testing
+
+import com.mapbox.api.directions.v5.models.DirectionsRoute
+import com.mapbox.navigation.testing.FileUtils
+
+object TestingUtil {
+    fun loadRoute(routeFileName: String): DirectionsRoute {
+        val routeAsJson = FileUtils.loadJsonFixture(routeFileName)
+        return DirectionsRoute.fromJson(routeAsJson)
+    }
+}
